Move type to messaging interfaces
