+++ rs/messaging/src/scheduling/valid_set_rule.rs
@@ -2,11 +2,14 @@ use ic_base_types::NumBytes;
 use ic_constants::{INGRESS_HISTORY_MAX_MESSAGES, SMALL_APP_SUBNET_MAX_SIZE};
 use ic_cycles_account_manager::{CyclesAccountManager, IngressInductionCost};
 use ic_error_types::{ErrorCode, UserError};
-use ic_interfaces::execution_environment::{
-    IngressHistoryWriter, IngressInductionError, LABEL_VALUE_CANISTER_METHOD_NOT_FOUND,
-    LABEL_VALUE_CANISTER_NOT_FOUND, LABEL_VALUE_CANISTER_OUT_OF_CYCLES,
-    LABEL_VALUE_CANISTER_STOPPED, LABEL_VALUE_CANISTER_STOPPING, LABEL_VALUE_INGRESS_HISTORY_FULL,
-    LABEL_VALUE_INVALID_MANAGEMENT_PAYLOAD,
+use ic_interfaces::{
+    execution_environment::IngressHistoryWriter,
+    messaging::{
+        IngressInductionError, LABEL_VALUE_CANISTER_METHOD_NOT_FOUND,
+        LABEL_VALUE_CANISTER_NOT_FOUND, LABEL_VALUE_CANISTER_OUT_OF_CYCLES,
+        LABEL_VALUE_CANISTER_STOPPED, LABEL_VALUE_CANISTER_STOPPING,
+        LABEL_VALUE_INGRESS_HISTORY_FULL, LABEL_VALUE_INVALID_MANAGEMENT_PAYLOAD,
+    },
 };
 use ic_logger::{debug, error, trace, ReplicaLogger};
 use ic_management_canister_types::CanisterStatusType;
